Changed conversion of Field * to long instead of int(to work with 64 bit architectures)
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@996 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/PythonMethods.cpp
@@ -1018,7 +1018,7 @@ call the base class __init__ function." );
 
       Py_DECREF( py_field_ptr );
       
-      return PyInt_FromLong( (int)field_ptr );
+      return PyInt_FromLong( (long)field_ptr );
       Py_INCREF(Py_None);
       return Py_None; 
     }
